producer/requirements.txt
pika==1.3.0
openai-whisper==20230314